ListViewControls can display a message when empty
Implemented as changing the display of an element linked to the object to block or none. By default it's a blank div not attached to the DOM, but it can be overridden by code by setting emptyMessage to a DOM element.

Also fixed a minor bug where the select all box was not updating when rows were added.