Controls are completely HTML/JS
Player no longer uses browser's audio controls, as all important functionality is provided by the player now. Aspects of the player control specific to a the implementation are controlled from a new playercontrols.js module. This will allow different players to be implemented later without touching the other frontend code, such as an SVG or a canvas-based player.