List elements are restricted to window size
Lists scroll seperately from other elements in the pane, including table headers. I tried a pure CSS solution, but there does not seem to be a way to get elements to fill available vertical space. Instead a Javascript function is called whenever the window is resized to fix the element heights.