Implemented scrolling list views
ListViewControls can now scroll apart from the rest of the page if a height for it is set. Furthermore, the header stays put while the content scrolls. It was necessary to split the control into two tables to it working right.

Rudimentary column widths have also been implemented, though they are all hard-coded right now.

Library and Queue are now using the new code and the page no longer scrolls when too much data is added.