Removed scan functionality from frontend
I realize I just moved it to it's "official" location in the last update, but I realized earlier today there's no good reason for the scanner to be on the web site, now that it can be run from the command line. If you really need to update the database remotely, SSH in and run it through manage.py.