Refactored backend
The multiple-API system was getting too complicated and trying to crowbar it onto Django was becoming increasingly unreliable. As there was only one package for it written, and no way properly defined for others, it has been removed for now. I will consider reimplementing a system for this once the codebase matures and the requirements are better known.