Adjusted audio source code
Audio source is determined by Javascript, not by multiple source elements in the DOM. This allows the best source to be used more reliably (e.g. don't use transcodes if we can help it) and to be more compatible in general. As it turns out, the HTML5 specs recommend not changing <source> tags dynamically. Now the code no longer even uses them.