Initial version of the manual
The very first version of the software's documentation, written in Sphinx. Currently all it has is apidoc sections; I will be writing more detailed documentation as time goes on.

Also first seen in this docs is the probable new name for the software, Cadence. I haven't committed to using it  yes, but I'm feeling it out right now.