Overhauled the transcoder system
The biggest change is that files in AUDIO_ROOT are no longer expected to be directly served now. Instead, the add_source_file() method symlinks to the original files and everything is served from the transcodes directory.

Another change is the removal of the skeleton code for muliple source files. This was deemed an unnecessary complexity that didn't really add any useful functionality while turning the source code into an unreadable mess, while also making it difficult to tell where your metadata is coming from.